+++ src/dktool.hpp
@@ -14,9 +14,9 @@ int dk_listen(SOCKET sk_fd);
 /// accept wrapper
 SOCKET dk_accept(SOCKET sk_fd, sockaddr_in *p_server_addr);
 /// socket read wrapper
-int dk_read(SOCKET fd,void *buffer, size_t n);
+ssize_t dk_read(SOCKET fd,void *buffer, size_t n);
 /// socket write wrapper
-int dk_write(SOCKET fd,void *buffer, size_t n);
+ssize_t dk_write(SOCKET fd,void *buffer, size_t n);
 /// socket connect weapper
 int dk_connect(SOCKET socket, const struct sockaddr *address,socklen_t address_len);
 /// start deamon process
